YAML-related functions

DFTraMO.dftramo_run — Function

dftramo_run(filename::AbstractString, software::AbstractString="vasp")

Automatically runs DFTraMO from a configuration yaml file.

DFTraMO.parse_energy — Method

parse_energy(software::AbstractString, energy::AbstractString) -> n::Float64

Parse energy values of the input yaml files and returns energies in Ha.

DFTraMO.parse_energy — Method

parse_energy(software::AbstractString, energy::Real) -> n::Float64

Parse energy values of the input yaml files and returns energies in Ha.

DFTraMO.parse_runs — Method

parse_runs(runs::Vector{Dict{Any, Any}}) -> runlist::Vector{RunInfo}

Parse the run section of the input yaml file and returns Vector{RunInfo}.

DFTraMO.parse_sites — Method

parse_sites(sites::AbstractVector{<:AbstractString}) -> site_final::Vector{Int}

Parses a portion of the "sites" lines in the yaml file to return a Vector{Int} with valid indices for targets. e.g.,

julia> parse_sites(["1:3", "3", "18:2:20"])
5-element Vector{Int64}:
  1
  2
  3
 18
 20

DFTraMO.read_run_yaml — Function

read_run_yaml(run_name::AbstractString)

Loads options for the run specified by a file.