YAML-related functions

DFTraMO.dftramo_run — Method

dftramo_run(filename::AbstractString)

Automatically runs DFTraMO from a configuration yaml file.

DFTraMO.parse_energy — Method

parse_energy(software::AbstractString, energy::AbstractString) -> n::Float64

Parse energy values of the input yaml files and returns energies in Ha.

DFTraMO.parse_energy — Method

parse_energy(software::AbstractString, energy::Real) -> n::Float64

Parse energy values of the input yaml files and returns energies in Ha.

DFTraMO.parse_runs — Method

parse_runs(runs::Vector{Dict{Any, Any}}) -> runlist::Vector{RunInfo}

Parse the run section of the input yaml file and returns Vector{RunInfo}.

DFTraMO.parse_sites — Method

parse_sites(sites::AbstractVector{<:AbstractString}) -> site_final::Vector{Int}

Parses a portion of the "sites" lines in the yaml file to return a Vector{Int} with valid indices for targets. e.g.,

julia> parse_sites(["1:3", "3", "18:2:20"])
5-element Vector{Int64}:
  1
  2
  3
 18
 20

DFTraMO.parse_yaml_energy — Method

DFTraMO.parse_yaml_energy(x, i::InputOrigin)

Converts a string or number representing an energy input, possibly with units given, into an energy in hartrees usable by DFTraMO. Units are not case-sensitive, and abbreviations are accepted as well as full unit names.

DFT-raMO will automatically assume quantities without explicit unit specification match the units of the software package which generated the input. If the software package is not specified or supported, DFT-raMO will assume Hartree atomic units.

DFTraMO.parse_yaml_length — Method

DFTraMO.parse_yaml_length(x, i::InputOrigin)

Converts a string or number representing an length input, possibly with units given, into a length in bohr usable by DFTraMO. Units are not case-sensitive, and abbreviations are accepted as well as full unit names.

Currently supported units are bohr (bohr, a0, or au), angstrom (Å, ang, or angstrom), nanometer (nm), and picometer (pm).

DFT-raMO will automatically assume quantities without explicit unit specification match the units of the software package which generated the input. If the software package is not specified or supported, DFT-raMO will assume Hartree atomic units.

DFTraMO.read_yaml — Method

DFTraMO.read_yaml(file)

Reimplementation of DFTraMO.read_run_yaml.